Use Format_ARGB32_Premultiplied instead of RGBA8888 for Qt backends.
I missed the fact that support for the latter was only added in Qt5.2
and is thus unavailable in Qt4.

Move the supporting helper functions from backend_cairo to cbook.
